// dist/node/community/schema.d.ts
export declare const CommunityChallengeSettingSchema: z.ZodObject<{
    path: z.ZodOptional<z.ZodString>;
    name: z.ZodOptional<z.ZodString>;
    options: z.ZodOptional<z.ZodRecord<z.ZodString, z.ZodString>>;
    exclude: z.ZodOptional<z.ZodArray<z.ZodObject<{
        community: z.ZodOptional<z.ZodObject<{
            addresses: z.ZodArray<z.ZodString>;
            maxCommentCids: z.ZodNumber;
            postScore: z.ZodOptional<z.ZodNumber>;
            replyScore: z.ZodOptional<z.ZodNumber>;
            firstCommentTimestamp: z.ZodOptional<z.ZodNumber>;
        }, z.core.$strict>>;
        postScore: z.ZodOptional<z.ZodNumber>;
        replyScore: z.ZodOptional<z.ZodNumber>;
        postCount: z.ZodOptional<z.ZodNumber>;
        replyCount: z.ZodOptional<z.ZodNumber>;
        firstCommentTimestamp: z.ZodOptional<z.ZodNumber>;
        challenges: z.ZodOptional<z.ZodArray<z.ZodNumber>>;
        role: z.ZodOptional<z.ZodArray<z.ZodUnion<[z.ZodEnum<{
            owner: "owner";
            admin: "admin";
            moderator: "moderator";
        }>, z.ZodString]>>>;
        address: z.ZodOptional<z.ZodArray<z.ZodString>>;
        rateLimit: z.ZodOptional<z.ZodNumber>;
        rateLimitChallengeSuccess: z.ZodOptional<z.ZodBoolean>;
        publicationType: z.ZodOptional<z.ZodObject<{
            post: z.ZodOptional<z.ZodBoolean>;
            reply: z.ZodOptional<z.ZodBoolean>;
            vote: z.ZodOptional<z.ZodBoolean>;
            commentEdit: z.ZodOptional<z.ZodBoolean>;
            commentModeration: z.ZodOptional<z.ZodBoolean>;
            communityEdit: z.ZodOptional<z.ZodBoolean>;
        }, z.core.$loose>>;
    }, z.core.$loose>>>;
    description: z.ZodOptional<z.ZodString>;
    pendingApproval: z.ZodOptional<z.ZodBoolean>;
}, z.core.$strict>;
// ...
export declare const CommunityChallengeSchema: z.ZodObject<{
    exclude: z.ZodOptional<z.ZodArray<z.ZodObject<{
        community: z.ZodOptional<z.ZodObject<{
            addresses: z.ZodArray<z.ZodString>;
            maxCommentCids: z.ZodNumber;
            postScore: z.ZodOptional<z.ZodNumber>;
            replyScore: z.ZodOptional<z.ZodNumber>;
            firstCommentTimestamp: z.ZodOptional<z.ZodNumber>;
        }, z.core.$strict>>;
        postScore: z.ZodOptional<z.ZodNumber>;
        replyScore: z.ZodOptional<z.ZodNumber>;
        postCount: z.ZodOptional<z.ZodNumber>;
        replyCount: z.ZodOptional<z.ZodNumber>;
        firstCommentTimestamp: z.ZodOptional<z.ZodNumber>;
        challenges: z.ZodOptional<z.ZodArray<z.ZodNumber>>;
        role: z.ZodOptional<z.ZodArray<z.ZodUnion<[z.ZodEnum<{
            owner: "owner";
            admin: "admin";
            moderator: "moderator";
        }>, z.ZodString]>>>;
        address: z.ZodOptional<z.ZodArray<z.ZodString>>;
        rateLimit: z.ZodOptional<z.ZodNumber>;
        rateLimitChallengeSuccess: z.ZodOptional<z.ZodBoolean>;
        publicationType: z.ZodOptional<z.ZodObject<{
            post: z.ZodOptional<z.ZodBoolean>;
            reply: z.ZodOptional<z.ZodBoolean>;
            vote: z.ZodOptional<z.ZodBoolean>;
            commentEdit: z.ZodOptional<z.ZodBoolean>;
            commentModeration: z.ZodOptional<z.ZodBoolean>;
            communityEdit: z.ZodOptional<z.ZodBoolean>;
        }, z.core.$loose>>;
    }, z.core.$loose>>>;
    description: z.ZodOptional<z.ZodString>;
    challenge: z.ZodOptional<z.ZodString>;
    type: z.ZodString;
    caseInsensitive: z.ZodOptional<z.ZodBoolean>;
    pendingApproval: z.ZodOptional<z.ZodBoolean>;
}, z.core.$loose>;